+++ fuzz/fuzz_targets/fuzz_arithmetic.rs
@@ -4,8 +4,8 @@ use std::str;
 
 use winnow::prelude::*;
 use winnow::{
-    branch::alt,
     character::{digit1 as digit, space0 as space},
+    combinator::alt,
     combinator::{delimited, terminated},
     multi::fold_many0,
 };
