+++ src/_tutorial/chapter_4.rs
@@ -47,7 +47,7 @@
 //! # use winnow::IResult;
 //! # use winnow::Parser;
 //! # use winnow::token::take_while1;
-//! use winnow::branch::dispatch;
+//! use winnow::combinator::dispatch;
 //! use winnow::token::take;
 //! use winnow::combinator::fail;
 //!
